+++ src/dashboard/components/Dashboard/Catalog/Products/Editor/ProductEditorWrapper.tsx
@@ -12,7 +12,7 @@ import { ReactElement, useEffect, useReducer, useState } from "react";
 import RootLayout from "@/pages/layout";
 // components
 import DashboardContentWithSaveFooter from "@/components/Dashboard/Generic/EditableContent";
-import TopLineWithReturn from "@/components/Dashboard/Catalog/Products/TopLineWithReturn";
+import TopLineWithReturn from "@/components/Dashboard/Generic/TopLineWithReturn";
 import ProductVariantsEditor from "@/components/Dashboard/Catalog/Products/Editor/Product/ProductVariantsEditor";
 import ProductMediaEditor from "@/components/Dashboard/Catalog/Products/Editor/Product/ProductMediaEditor";
 import ProductVariantPricesEditor from "@/components/Dashboard/Catalog/Products/Editor/Product/ProductVariantPricesEditor";
