+++ src/dashboard/pages/dashboard/catalog/product-types/[id].tsx
@@ -9,7 +9,7 @@ import { ReactElement, useEffect, useState } from "react";
 import RootLayout from "@/pages/layout";
 // components
 import DashboardContentWithSaveFooter from "@/components/Dashboard/Generic/EditableContent";
-import TopLineWithReturn from "@/components/Dashboard/Catalog/Products/TopLineWithReturn";
+import TopLineWithReturn from "@/components/Dashboard/Generic/TopLineWithReturn";
 import ProductTypeGeneralInformation from "@/components/Dashboard/Catalog/ProducType/ProductTypeGeneralInformation";
 import ProductTypeAllowedAttribtuesSelect from "@/components/Dashboard/Catalog/ProducType/ProductTypeAllowedAttribtuesSelect";
 // mui
